examples/agno-support-agent/pyproject.toml
[project]
name = "majordomo-agno-example"
version = "0.1.0"
description = "Agno agent examples with Majordomo Gateway integration"
requires-python = ">=3.12"
dependencies = [
    "majordomo-frameworks[agno]",
    "anthropic>=0.77.1",
    "google-genai>=1.62.0",
    "python-dotenv>=1.0.0",
    "click>=8.3.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.9.0",
]

[tool.uv.sources]
majordomo-frameworks = { path = "../.." }
